requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.36
botocore==1.38.36
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mysqlclient==2.2.7
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
